+++ media-libs/opencv/opencv-4.9.0-r1.ebuild
@@ -919,6 +919,7 @@ PATCHES=(
 	# "${FILESDIR}/${PN}_contrib-${PV}-rgbd.patch"
 	# "${FILESDIR}/${PN}_contrib-4.8.1-NVIDIAOpticalFlowSDK-2.0.tar.gz.patch"
 	"${FILESDIR}/${PN}-4.9.0-openvx-paths.patch" # oiledmachine-overlay change
+	"${FILESDIR}/${PN}-4.10.0-vulkan-libdirs.patch"
 )
 
 cuda_get_cuda_compiler() {
@@ -1001,7 +1002,9 @@ src_prepare() {
 	# Remove bundled stuff
 	mkdir -p "3rdparty.orig" || die
 	mv "3rdparty/flatbuffers" "3rdparty.orig" || die
-	use carotene && mv "3rdparty/carotene" "3rdparty.orig" || die
+	if use carotene ; then
+		mv "3rdparty/carotene" "3rdparty.orig" || die
+	fi
 	rm -r "3rdparty" || die "Removing 3rd party components failed"
 	if use system-flatbuffers ; then
 		rm -rf "3rdparty.orig" || die
@@ -1500,6 +1503,10 @@ eerror "OpenVINO is not supported for ${ARCH}"
 		export PKG_CONFIG_PATH="/usr/lib/ffmpeg/58.60.60/$(get_libdir)/pkgconfig:${PKG_CONFIG_PATH}"
 	elif use ffmpeg && has_version "media-video/ffmpeg:56.58.58" ; then # 4.x
 		export PKG_CONFIG_PATH="/usr/lib/ffmpeg/56.58.58/$(get_libdir)/pkgconfig:${PKG_CONFIG_PATH}"
+	fi
+
+	if use vulkan ; then
+		export VULKAN_SDK="/usr"
 	fi
 
 	if multilib_is_native_abi && use python ; then
